+++ crates/cli/src/commands/tool/mod.rs
@@ -1,11 +1,11 @@
 mod add;
 mod info;
 mod list;
-mod plugins;
+mod list_plugins;
 mod remove;
 
 pub use add::*;
 pub use info::*;
 pub use list::*;
-pub use plugins::*;
+pub use list_plugins::*;
 pub use remove::*;